+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - Fix return type of `String.charCodeAt`. https://github.com/rescript-lang/rescript/pull/7864
 - Remove support of JSX children spread. https://github.com/rescript-lang/rescript/pull/7869
+- Have `String.charCodeAt` return `option<int>`; add `String.charCodeAtUnsafe`. https://github.com/rescript-lang/rescript/pull/7877
 
 #### :eyeglasses: Spec Compliance